Yay Parsing
This should be near a finished idea. - not the finished product, but as
far as parsing it, without actually doing the DBMS, this should be
almost complete. there are still a few functions to do. but it is
working nicely. I am going to upload the .db file I have been using in a
second. As Ross pointed out earlier, make sure to put it in your project
file before compiling.