Testing for correct syntax
new broken DB for testing if lines are correct or not, error testing for
the parser. what we have now would be fine to turn in I think, but it
needs a few more tweaks to be solid. dont forget to change table.txt to
table.db when moving it into the project file.