stuck and confused
got stuck and confused on what RELATION and project means, so I kinda
did projection. for renaming, there seems to be a bug I can't figure
out. For update and delete, I'm having a hard time understanding the
parser and DBMS on how to do it. If maybe the code master (Lance) wants
to give it a try. Sorry I'm a bit stumped by your code.